// src/app/model/EmployeeModel.ts
export class EmployeeModel {

    [key: string]: any;
    Id: any;
    FullName: any;
    Email: any;
    PhoneNo: any;
    Address: any;
    Pincode: any;
    DeptID: any;
    DesigID: any;
    City: any;
    state: any;
    Gender: any;
    Qualification: any;
    Citizen: any;
    PanNO: any;
    Exp: any;
    HireDate: any;
    DOB: any;
    language: any;
    skills: any;
    QualificationMarks: any;
    Nominee: any;
    IsActive: boolean | null;
    MarTialstatus: any;
}